Add inline extern definitions of table_get functions
These functions are declared in oslmic.h as inline functions, with a
body to allow inlining. In C++, when a function is used but not inlined
everywhere, it is output into every compilation unit where it is used
like that, with the linker then discarding all but one definition.

However, in C, a definition is only output when a function is declared
as extern inline. In practice, this means the function body stays in the
header file, but an additional extern inline declaration is inserted
into a single source file to output the function definition into that
compilation unit (regardless of whether it is actually used in that
compilation unit).

These extern inline declarations were missing for the tablet_get
functions introduced in commit acb75c5 (Use macros for all constant
table definitions and lookups). This was not a problem initially,
probably because all calls were inlined. In the Arduino AVR core version
1.8, the compiler was upgraded to gcc 7, which started producing
undefined references, possibly because some calls are no longer inlined.

This commit adds the extern inline declarations for these functions.

This fixes #243